% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_performance_class.R
\docType{methods}
\name{bin_,Binary_Performance,Continuous-method}
\alias{bin_,Binary_Performance,Continuous-method}
\title{Bin a Continuous object using Binary_Performance}
\usage{
\S4method{bin_}{Binary_Performance,Continuous}(.self, b, min.iv = 0.01,
  min.cnt = 10, min.res = 0, max.bin = 10, mono = 0,
  exceptions = numeric(0))
}
\arguments{
\item{.self}{a Binary_Performance object}

\item{b}{a Continuous bin object}

\item{min.iv}{minimum information value to split}

\item{min.cnt}{minimum number of observations per bin}

\item{min.res}{minimum number of response counts per bin}

\item{max.bin}{maximum number of resulting bins}

\item{mono}{montoncity constraint of discretization process}

\item{exceptions}{numeric vector of values to exclude from discretization#'}
}
\value{
modifies the Bin object in place
}
\description{
Bin a Continuous object using Binary_Performance
}

