% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_performance_class.R
\docType{methods}
\name{bin_,Binary_Performance,Discrete-method}
\alias{bin_,Binary_Performance,Discrete-method}
\title{Bin a Discrete object using Binary_Performance}
\usage{
\S4method{bin_}{Binary_Performance,Discrete}(.self, b)
}
\arguments{
\item{.self}{A Binary_Performance object}

\item{b}{A Discrete bin object}
}
\value{
modifies the Bin object in place
}
\description{
Bin a Discrete object using Binary_Performance
}

