% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_class.R
\name{bin_}
\alias{bin_}
\title{generic method for bin_}
\usage{
bin_(.self, b, ...)
}
\arguments{
\item{.self}{the Performance object calling bin_}

\item{b}{a Bin object}

\item{...}{further arguments passed on to bin_ methods}
}
\description{
generic method for bin_
}

