% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_class.R
\name{Bin_as.matrix}
\alias{Bin_as.matrix}
\title{Return matrix summarizing Bin object}
\value{
\code{matrix} of summarized bin data
}
\description{
Return matrix summarizing Bin object
}
\details{
\code{as.matrix} summarizes the \code{\link{Performance-class}}
object within each level of the Bin object. As such, the summarization
process must be described by implementing a summarize method for the
Performance object.
}

