% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_class.R
\name{Bin_bin}
\alias{Bin_bin}
\title{Discretize and summarize variables}
\arguments{
\item{...}{arguments passed on to the Performance object's \code{bin}
implementation. This varies depending on the performance type.}
}
\value{
\code{bin} creates a Transform object and adds it to the Bin object.
}
\description{
The bin function is the entry point for discretizing and
summarizing variables by some arbitray performance metric. Each Performance
object must implement a bin method for numeric and factor variables.
Currently only binary performance is supported.
}
