% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_class.R
\docType{class}
\name{Bin_Class}
\alias{Bin_Class}
\alias{Bin}
\title{Bin reference class generator}
\description{
Bin object generator class used to wrap binned variables.
}
\section{Fields}{

\describe{
\item{\code{x}}{numeric or factor vector to be discretized or summarized}

\item{\code{name}}{name of the variable}

\item{\code{perf}}{\code{\link{Performance}} object used to discretize and summarize
\code{x}. Currently only binary performance supported.}

\item{\code{tf}}{\code{\link{Transform}} object containing information for
discretizing and summarizing \code{x}.}

\item{\code{history}}{\code{list} of \code{Transform} objects. One for every
operation on the \code{Bin} object.}

\item{\code{args}}{saved arguments that were used to call the bin method.}
}}

\section{Methods}{

\describe{
\item{\code{expand(...)}}{Updated Bin after inherited expand}

\item{\code{factorize(newdata = .self$x)}}{Return list of filters for exceptions, missing, and normal values}
}}

