% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_class.R
\name{Bin_collapse}
\alias{Bin_collapse}
\title{Collapse levels of a Bin object}
\arguments{
\item{i}{numeric vector of bin levels to collapse.}
}
\value{
modifies the transform object in place.
}
\description{
Collapse levels of a Bin object
}
\seealso{
Continuous_collapse Discrete_collapse
}

