% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_class.R
\name{Bin_expand}
\alias{Bin_expand}
\title{Expand a level of a Bin into multiple new levels}
\arguments{
\item{i}{numeric vector of length 1 indiicating bin level to expand.}
}
\value{
modifies the transform object in place.
}
\description{
Expand a level of a Bin into multiple new levels
}
\details{
All of the collapsed levels will be expanded.
}
\seealso{
Continuous_expand Discrete_expand
}

