% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_class.R
\name{Bin_factorize}
\alias{Bin_factorize}
\title{Factrorize superclass method}
\arguments{
\item{newdata}{vector on which to apply the transformation. Defaults to the
\code{x} field of the Bin object}
}
\value{
\code{list} with three fields. See details.
}
\description{
Factrorize superclass method
}
\details{
\code{factorize} returns a list with three fields:
\itemize{
 \item{normal }{ logical vector indicating non-missing, non-exceptoin values}
 \item{exception }{ logical vector indicating exception values}
 \item{missing }{ logical vector indicating missing values}
}
}
