% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{Bin_minus}
\alias{-,Bin,numeric-method}
\alias{Bin_minus}
\title{Collapse Bins}
\usage{
\S4method{-}{Bin,numeric}(e1, e2)
}
\arguments{
\item{e1}{A Continuous or Discrete bin}

\item{e2}{the levels to collapse}
}
\description{
Collapse Bins
}

