% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_class.R
\name{Bin_neutralize}
\alias{Bin_neutralize}
\title{Set performance value to zero}
\arguments{
\item{i}{numeric vector of bin indices to neutralize.}
}
\description{
\code{neutralize} sets the performance substitution value of
the requested indices to zero. This has the effect of neither adding nor
subtracting points in the final score.
}

