% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{Bin_not_equal}
\alias{!=,Bin,numeric-method}
\alias{Bin_not_equal}
\title{Neutralize Bins}
\usage{
\S4method{!=}{Bin,numeric}(e1, e2)
}
\arguments{
\item{e1}{A Continuous or Discrete bin}

\item{e2}{the levels to neutralize}
}
\description{
Neutralize Bins
}

