% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{Bin_plus}
\alias{Bin_plus}
\title{Expand Bins}
\usage{
\S4method{+}{Bin,numeric}(e1, e2)
}
\arguments{
\item{e1}{A Continuous or Discrete bin}

\item{e2}{The levels to expand}
}
\description{
Expand Bins
}
