% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_class.R
\name{Bin_predict_sparse}
\alias{Bin_predict_sparse}
\title{Substitute weight-of-evidence for the input \code{x} values}
\arguments{
\item{newdata}{vector of the same type as \code{x} for which to substitute
the bin weight-of-evidence values.}
}
\description{
Substitute weight-of-evidence for the input \code{x} values
}
