% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_class.R
\name{Bin_set_equal}
\alias{Bin_set_equal}
\title{Set one level equal to another}
\arguments{
\item{i1}{index of target bin}

\item{i2}{index of source bin}
}
\description{
\code{set_equal} sets the performance summary value of \code{i1}
equal to that of \code{i2}. This can be used to force two bins to have the
same substituted value.
}
