% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_class.R
\name{Bin_set_overrides}
\alias{Bin_set_overrides}
\title{Override values of subst/pred with bin-level coefficients}
\arguments{
\item{i1}{overrides of all values}
}
\description{
Override values of subst/pred with bin-level coefficients
}
