% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_class.R
\name{Bin_show}
\alias{Bin_show}
\title{Print representation of Bin object}
\arguments{
\item{...}{optional arguments passed on to the \link{\code{print}} function.}
}
\description{
Print representation of Bin object
}
