% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_class.R
\name{Bin_summary}
\alias{Bin_summary}
\title{Return summary information for Bin based on associated performance object}
\description{
Return summary information for Bin based on associated performance object
}

