% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_class.R
\name{Bin_update}
\alias{Bin_update}
\title{Update the Bin object after performing requested operations}
\description{
The update method is called any time the user makes changes to
the Bin object. The transform object is changed and the history list is
updated.
}

