% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_performance_class.R
\name{Binary_Performance_bin}
\alias{Binary_Performance_bin}
\title{Subsequent call from the bin function passed to Binary_Performance object}
\description{
This bin function should not be directly called by the user.
The Classing bin function is subsequently called from the
\code{\link{Bin_bin}} wrapper function.
}
