% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_performance_class.R
\docType{class}
\name{Binary_Performance_Class}
\alias{Binary_Performance_Class}
\alias{Binary_Performance}
\title{Bin reference class generator}
\description{
Binary_Performance object generator class that implements
the Performance class methods
}
\section{Fields}{

\describe{
\item{\code{ones}}{sum of \code{y == 0}}

\item{\code{zeros}}{sum of \code{y == 1}}
}}


