% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_performance_class.R
\name{Binary_Performance_get_excel_table}
\alias{Binary_Performance_get_excel_table}
\title{Return an object that tells openxlsx how to print it to Excel}
\arguments{
\item{b}{Bin object to update}
}
\value{
Returns a \code{openxlsx-table} object that \code{binnrtools} can
use to format a bin object for Excel
}
\description{
Return an object that tells openxlsx how to print it to Excel
}
