% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_performance_class.R
\name{Binary_Performance_plot}
\alias{Binary_Performance_plot}
\title{Plot method implementation for Binary_Performance}
\arguments{
\item{b}{Bin object to update}
}
\description{
plot displays a horizontal bar chart showing summarized
performance information. The large bars represent the observed WoE within
each bin. The smaller bars within represent the WoE that will be substituted
during prediction. Record count percentages are shown on the right side of
the plot while a sequential index used for referencing bin levels is printed
along the left.
}
