% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_performance_class.R
\name{Binary_Performance_sort_value}
\alias{Binary_Performance_sort_value}
\title{Sort_value method implementation for Binary_Performance}
\arguments{
\item{b}{Bin object to update}
}
\value{
the value to use for sorting bins using Binary_Performance
}
\description{
sort_value returns the information value for the requested Bin
}

