% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_performance_class.R
\name{Binary_Performance_summarize}
\alias{Binary_Performance_summarize}
\title{Summarize method implementation for Binary_Performance}
\arguments{
\item{x}{discretized independent variable as a factor}

\item{y}{response variable}

\item{w}{weight variable}
}
\value{
a matrix of summary information for every level in the factor x
}
\description{
this function summarizes the relationship between the
independent and response variables. The inherited Performance class demands
that it be implemented.
}
