% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_performance_class.R
\name{Binary_Performance_summary}
\alias{Binary_Performance_summary}
\title{Summary method implementation for Binary_Performance}
\arguments{
\item{tf}{Transform object to summarize}
}
\value{
a name vector of summary information
}
\description{
summary returns the appropriate pieces from the Transform object
to summarize the relationship in a conciese one-line entry.
}

