% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_performance_class.R
\name{Binary_Performance_update}
\alias{Binary_Performance_update}
\title{Update method implementation for Binary_Performance}
\arguments{
\item{b}{Bin object to update}
}
\value{
a list of matrices with summarized information.
}
\description{
update is called after every applicable bin operation that
modifies the Transform object. The result of calling update is used for
displaying the Bin matrix as well as plotting the Bin object.
}
