% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binnr-package.R
\docType{package}
\name{binnr-package}
\alias{binnr}
\alias{binnr-package}
\title{binnr: Biologically Informed Neural Networks as Graph Neural Networks}
\description{
Fits biologically informed neural networks (BINNs) whose architecture follows a pathway knowledge graph such as Reactome. Models are implemented as message-passing graph neural networks on a directed acyclic graph using 'torch', so that classic masked feedforward BINNs are a special case, and constraints such as scalar pathway nodes, fixed depth, dummy padding nodes and edge-specific weights can be relaxed. Supports multi-omics inputs, clinical covariates and classification, regression and Cox survival outcomes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datasciapps/binnr}
  \item Report bugs at \url{https://github.com/datasciapps/binnr/issues}
}

}
\author{
\strong{Maintainer}: David Selby \email{david_antony.selby@dfki.de} (\href{https://orcid.org/0000-0001-8026-5663}{ORCID}) (affiliation: Deutsches Forschungszentrum für Künstliche Intelligenz (DFKI))

Authors:
\itemize{
  \item David Selby \email{david_antony.selby@dfki.de} (\href{https://orcid.org/0000-0001-8026-5663}{ORCID}) (affiliation: Deutsches Forschungszentrum für Künstliche Intelligenz (DFKI))
  \item Daria Pavliuk \email{daria.pavliuk@dfki.de} (affiliation: DFKI)
  \item Gerrit Grossmann \email{gerrit.grossmann@dfki.de} (\href{https://orcid.org/0000-0002-4933-447X}{ORCID}) (affiliation: DFKI)
  \item Sebastian Vollmer \email{sebastian.vollmer@dfki.de} (\href{https://orcid.org/0000-0003-2831-1401}{ORCID}) (affiliation: DFKI)
}

}
\keyword{internal}
