% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{bins.to.csv}
\alias{bins.to.csv}
\title{Bins to CSV}
\usage{
bins.to.csv(bins, file = NULL)
}
\arguments{
\item{bins}{The \code{bin.object} to export}

\item{file}{The fully qualified path to write the bins}
}
\description{
Export bins to csv file that can be opened in Excel
}

