% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_class.R
\docType{class}
\name{Classing-class}
\alias{Classing-class}
\alias{Classing}
\title{Class Classing}
\description{
Classing class that wraps a data.frame and prepares it for
scorecard modeling.
}
\section{Fields}{

\describe{
\item{\code{variables}}{list of binned variables}

\item{\code{vnames}}{character vector of binned variable names}

\item{\code{performance}}{Performance object with \code{y} and \code{w} values and
implementation of the Performance object interface.}

\item{\code{dropped}}{character vector of variable names that are flagged as
dropped}
}}


