% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_class.R
\name{Classing_bin}
\alias{Classing_bin}
\title{Subsequent call from the bin function passed to Classing object}
\description{
This bin function should not be directly called by the user.
The Classing bin function is subsequently called from the
\code{\link{bin}} wrapper function.
}
