% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_class.R
\name{Classing_cluster}
\alias{Classing_cluster}
\title{Cluster variables by correlation}
\arguments{
\item{keep}{logical indicating whether to include dropped variables in the
correlation cluster analysis.}

\item{bag.fraction}{sample the development data without replacement to
cluster more quickly. The bag.fraction is multiplied by the number of
observations.}
}
\value{
a classing_cluster object
}
\description{
Cluster variables by correlation
}
\details{
the cluster function first performs weight-of-evidence substitution.
Cluster returns a classing_cluster object which is a list containing
two fields: correlation & cluster. The first is a correlation matrix for all
of the variables in the classing. The second is an hclust object which is
result of hierarchical clustering of the correlation matrix.
}

