% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_class.R
\name{Classing_combine}
\alias{Classing_combine}
\title{Add variables from one classing to another}
\arguments{
\item{other}{an object of class \code{\link{Classing_Class}} that will be
incorporated into the calling method's Classing object}
}
\value{
an updated Classing object
}
\description{
Add variables from one classing to another
}
\details{
\code{combine} adds a Classing object's variables to the caller
object. The 'other' Classing object must not have any of the same variables.
It must also use the same performance as the calling method's Classing
object.
}
