% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_class.R
\name{Classing_get_clusters}
\alias{Classing_get_clusters}
\title{Return a data.frame summarizing the variable clusters}
\arguments{
\item{cc}{classing_cluster object produced by \link{\code{Scorecard_cluster}}
method}

\item{corr}{minimum correlation coefficient threshold with which to group
variables}
}
\value{
a data.frame of variable clusters and information value
}
\description{
Return a data.frame summarizing the variable clusters
}
