% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_class.R
\name{Classing_prune_clusters}
\alias{Classing_prune_clusters}
\title{Prune clusters keeping only the most informative variables}
\arguments{
\item{cc}{classing_cluster object produced by \code{\link{Classing_cluster}}
method}

\item{corr}{minimum correlation coefficient threshold with which to group
variables}

\item{n}{number of variables to keep from each cluster exceeding the
correlation threshold. The n variables with the highest information value
are retained. The remaining variables are returned as a character vector.}
}
\value{
a character vector of variables to drop
}
\description{
Prune clusters keeping only the most informative variables
}

