% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_class.R
\name{Classing_set_step}
\alias{Classing_set_step}
\title{Flag supplied variables as dropped}
\arguments{
\item{vars}{character vector of variables to drop}

\item{all}{logical indicating whether all variables should be dropped}
}
\description{
Flag supplied variables as dropped
}
