% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_class.R
\name{Classing_summary}
\alias{Classing_summary}
\title{Summarize the Classing object}
\value{
a matrix summarizing the independent variables using the Performance
object summary function
}
\description{
Summarize the Classing object
}

