% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare binnr models}
\usage{
compare(...)
}
\arguments{
\item{...}{Any number of \code{binnr.model} objects}
}
\value{
A \code{data.frame} of model contributions merged by variable name
and sorted by the last, passed model in descending order
}
\description{
Compare binnr models
}
\details{
\code{compare} can be used to assess the effects of dropping variables
from a \code{binnr.model} and refitting. Dropped variables will show up as 'NA'
in the re-fit model. Newly added variables will show up as 'NA' in the previous
model
}

