% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_class.R
\name{Continuous_collapse}
\alias{Continuous_collapse}
\title{Collapse adjacent levels of a Continuous bin object}
\arguments{
\item{i}{numeric vector of bin levels to collapse. Must be adjacent.}
}
\value{
modifies the transform object in place.
}
\description{
Collapse adjacent levels of a Continuous bin object
}
