% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_class.R
\name{Continuous_exceptions}
\alias{Continuous_exceptions}
\title{Set the exception values for Continuous bins}
\arguments{
\item{e}{numeric vector of exception values to withhold from binning}
}
\description{
Set the exception values for Continuous bins
}

