% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_class.R
\name{Continuous_expand}
\alias{Continuous_expand}
\title{a level of a Continuous bin into multiple new levels}
\arguments{
\item{i}{numeric vector of length 1 indiicating bin level to expand.}
}
\value{
modifies the transform object in place.
}
\description{
a level of a Continuous bin into multiple new levels
}
\details{
The requested level is divided into quintiles if possible. Duplicate
levels are removed.
}
