% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_class.R
\name{Continuous_fmt_numeric_cuts}
\alias{Continuous_fmt_numeric_cuts}
\title{Helper function to format Continuous bin labels}
\value{
character vector of bin labels
}
\description{
Helper function to format Continuous bin labels
}
