% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_class.R
\name{Continuous_predict}
\alias{Continuous_predict}
\title{Weight-of-Evidence subistitution for Continuous bins}
\arguments{
\item{newdata}{numeric vector to apply performance substition. Defaults to
data used to create the Continuous object.}
}
\value{
numeric variable with bin performance values substituted for
the inputs.
}
\description{
Weight-of-Evidence subistitution for Continuous bins
}
