% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_class.R
\name{Continuous_set_cutpoints}
\alias{Continuous_set_cutpoints}
\title{Explicity set bin boundaries for Continuous objects}
\arguments{
\item{cuts}{space-separated list of numeric cutpoints}
}
\description{
\code{set_equal} sets the performance summary value of \code{i1}
equal to that of \code{i2}. This can be used to force two bins to have the
same substituted value.
}
