% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_class.R
\docType{class}
\name{Discrete-class}
\alias{Discrete}
\alias{Discrete-class}
\title{Discrete class}
\description{
Discrete object generator class used to wrap binned factor variables.
}
\section{Methods}{

\describe{
\item{\code{expand(...)}}{Updated Bin after inherited expand}

\item{\code{factorize(newdata = .self$x)}}{Return list of filters for exceptions, missing, and normal values}
}}

