% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_class.R
\name{Discrete_collapse}
\alias{Discrete_collapse}
\title{Collapse levels of a Discrete bin object}
\arguments{
\item{i}{numeric vector of bin levels to collapse. Do not have to be
adjacent.}
}
\value{
modifies the transform object in place.
}
\description{
Collapse levels of a Discrete bin object
}

