% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_class.R
\name{Discrete_expand}
\alias{Discrete_expand}
\title{Expand a level of a Discrete bin into multiple new levels}
\arguments{
\item{i}{numeric vector of length 1 indiicating bin level to expand.}
}
\value{
modifies the transform object in place.
}
\description{
Expand a level of a Discrete bin into multiple new levels
}
\details{
All of the collapsed levels will be expanded.
}
