% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_class.R
\name{Discrete_factorize}
\alias{Discrete_factorize}
\title{Factorize for Discrete bins}
\arguments{
\item{newdata}{Factor vector on which to apply the transformation. Defaults
to the \code{x} field of the Discrete object}
}
\value{
\code{list} with two fields. See details.
}
\description{
Factorize for Discrete bins
}
\details{
\code{factorize} returns a list with two fields:
\itemize{
 \item{factor }{ Factor with the bin labels applied to \code{x}}
 \item{types }{ list of logical vectors for missing, exception, and normal}
}
}
