% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{fit binnr objects}
\usage{
fit(x, data = NULL, y = NULL, nfolds = 3, lower.limits = 0,
  upper.limits = 3, family = "binomial", alpha = 1, drop = F,
  fixed = F)
}
\arguments{
\item{x}{}

\item{data}{a \code{data.frame} with a names attribute or NULL}

\item{y}{a numeric vector the same length as the number of rows in \code{data}
or NULL}

\item{nfolds}{Number of cross-validation folds for \code{cv.glmnet}}

\item{lower.limits}{minimum value of coefficients in the glmnet model}

\item{upper.limits}{maximum value of coefficients in the glmnet model}

\item{family}{Link function family to pass on to \code{cv.glmnet}}

\item{alpha}{parameter passed to \code{cv.glmnet}. 0 == LASSO, 1 == Ridge}

\item{drop}{boolean indicator for whether to flag variables not in the model
as "dropped" which will prevent them from being fit in subsequent iterations.}

\item{x}{a \code{bin.list}, \code{binnr.model}, or \code{segmented} object}
}
\value{
a \code{binnr.model} or \code{segmented} object depending on
what was passed.
}
\description{
\code{fit} is the main engine for fitting binnr objects to data.
}
\details{
\code{fit} does a lot of heavy lifting for the user. It will apply
the weight of evidence transformations stored in the bins of the passed object.
If provided a dataset, it will apply these transformations on the dataset. If
not given a dataset, \code{fit} will use the data that was provided to create
the bins.

The transformed data is passed to \code{cv.glmnet} to fit a LASSO regression
model.
}

