% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R
\docType{class}
\name{Model-class}
\alias{Model-class}
\title{Model class definition}
\description{
The Model object contains all of the information describing a
fitted binnr Scorecard. This includes a list of the Transforms used for WoE
substitution, which variables entered the model, and some performance
metrics.
}
\section{Slots}{

\describe{
\item{\code{name}}{short model identifier}

\item{\code{description}}{short description of the model}

\item{\code{settings}}{saved list of settings used to bin and fit the model}

\item{\code{transforms}}{list of Bin Transforms used to fit the model}

\item{\code{dropped}}{character vector denoting which variables were dropped at time
of fit}

\item{\code{inmodel}}{character vector denoting which variables entered the model}

\item{\code{coefs}}{intercept and coefficients of the fitted model as a numeric
vector}

\item{\code{contribution}}{some measure of variable contribution to the model's
predictive power}

\item{\code{ks}}{statistic measuring the separation of the two performance classes}

\item{\code{fit}}{the result of the cv.glmnet function call}
}}

