% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_class.R
\name{neutralize}
\alias{neutralize}
\alias{neutralize_}
\title{Neutralize selected levels of Transform setting substitution to zero}
\usage{
neutralize_(tf, i)
}
\arguments{
\item{tf}{Transform object}

\item{i}{numeric vector of levels to neutralize}
}
\value{
new Transform object with updated overrides
}
\description{
Neutralize selected levels of Transform setting substitution to zero
}
