% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_class.R
\docType{class}
\name{Performance-class}
\alias{Performance-class}
\alias{Performance}
\title{Performnace reference class generator#'}
\description{
Performnace reference class generator#'
}
\section{Fields}{

\describe{
\item{\code{y}}{numeric or factor response variable}

\item{\code{w}}{numeric weight variable}
}}


