% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rc.report.R
\name{rc.table}
\alias{rc.table}
\title{Reason Code Table}
\usage{
rc.table(x, data, nret, seg, mode)
}
\arguments{
\item{x}{a \code{segmented} or \code{binnr.model} object}

\item{data}{a \code{data.frame} to predict rcs}

\item{nret}{number of reason code positions returned}

\item{seg}{segment variable if called on \code{segmented} object}
}
\value{
a data.frame or list of data.frames with detailed
reason code reporting statistics
}
\description{
Create a reason code distribution table from binnr objects
}
\details{
If called on a data.frame, rc.table returns a single
data.frame with detailed statistcs about the number and percentage
of returned and triggered reason codes. If called on a \code{segmented}
object, rc.table returns a list of data.frames - one for each level
found of the segment variable found in the data.
}

