% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sas.R
\name{sas}
\alias{sas}
\title{Create character vector of SAS code for printing}
\usage{
sas(x, ...)
}
\arguments{
\item{x}{a \code{bin} or \code{bin.list} object}
}
\value{
a character vector of SAS statements. Use \code{cat} to export
}
\description{
Create character vector of SAS code for printing
}

