% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard_class.R
\docType{class}
\name{Scorecard-class}
\alias{Scorecard}
\alias{Scorecard-class}
\title{Scorecard reference class generator}
\description{
Scorecard class that wraps a data.frame and prepares it for
scorecard modeling.
}
\section{Fields}{

\describe{
\item{\code{seed}}{saved random seed that is based on  \code{\link{Sys.time}}.}

\item{\code{models}}{list of fitted models for this Scorecard}

\item{\code{selected_model}}{name of last  selected that was loaded into the
transforms}

\item{\code{inmodel}}{character vector of variable names that are in the selected
model}
}}
\section{Methods}{

\describe{
\item{\code{add_model(mod)}}{Register a new Model object with the scorecard}

\item{\code{has_model(model)}}{Assert whether the Scorecard contains the requested model}
}}

