% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard_class.R
\name{Scorecard_bin}
\alias{Scorecard_bin}
\title{Subsequent call from the bin function passed to the Scorecard}
\description{
This bin function should not be directly called by the user.
The Scorecard bin function is subsequently called from the
\link{\code{bin}} wrapper function.
}
