% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard_class.R
\name{Scorecard_bootstrap}
\alias{Scorecard_bootstrap}
\title{Run bootstrap model fits to assess coefficient distributions.}
\arguments{
\item{times}{number of bootstrap samples to run}

\item{bag.fraction}{fraction of observations to sample for each bootstrap run}

\item{replace}{whether to sample with or without replacement}

\item{nfolds}{number of CV folds to run within each bootstrap fit}

\item{upper.limits}{maximum coffecient value for each fit}

\item{lower.limits}{minimum coffecient value for each fit}

\item{alpha}{mixing paramater between LASSO and Ridge regression. Default 1.}

\item{...}{other parameters passed on to cv.glmnet}
}
\value{
a list with two elements: pvals and coefficients. The former is a
vector indicating what proportion of bootstrap model fits each coefficient
returned a zero. The latter an nRuns x nVars matrix containing the
coefficients of each run.
}
\description{
Run bootstrap model fits to assess coefficient distributions.
}
\details{
True boostratp samples should be run with \code{bag.fraction=1} and
\code{replace=TRUE}.
}

