% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard_class.R
\name{Scorecard_compare}
\alias{Scorecard_compare}
\title{Compare multiple scorecards side-by-side}
\description{
Compare multiple scorecards side-by-side
}
\details{
calling adjust enters an interactive variable edit mode. Press "h"
for a list of commands.
}
