% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard_class.R
\name{Scorecard_fit}
\alias{Scorecard_fit}
\title{Fit a model to the current set of variable transforms}
\arguments{
\item{name}{brief model name as character}

\item{description}{character description describing the model}

\item{overwrite}{should model be overwriiten if it already exists?}

\item{newdata}{data.frame of independent variables. Default is to
use the binned data.}

\item{y}{target to fit the data to. Default is the y variable used
for discretization.}

\item{w}{optional weight variable}

\item{nfolds}{number of k-folds with which to select the optimal
lambda value}

\item{upper.limits}{maximum value of fitted coefficients}

\item{lower.limits}{minimimum value of fitted coefficients}

\item{alpha}{type of regularization. Default is alpha == 1 for LASSO
regression. Alpha of 0 is Ridge regression while anythin in between
is the elastic net mixture.}

\item{family}{response variable distribution. Default is "binomial".}

\item{...}{additional arguments passed on to cv.glmnet}
}
\description{
fits a regularized regressoion model using the glmnet
package.
}
\details{
the fit function first calls predict and substitutes the
weight-of-evidence for all predictor variables. It then passes this matrix
on to \code{\link[glmnet]{cv.glmnet}}. The coefficients of a binner model fit
are restricted to [0,3]. This ensures there are no sign flips in the
model coefficients and that the relationships observed on margin are
retained in the final model.
\code{\link{bin}} wrapper function.
}

