% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard_class.R
\name{Scorecard_gen_code_sas}
\alias{Scorecard_gen_code_sas}
\title{Generate SAS code for Scorecard object}
\arguments{
\item{pfx}{character prefix to prepend to variable names
by the Scorecard model object. Defaults to 1.}

\item{method}{method used for calculating the reference level for adverse
action codes. Three possible choices:
\itemize{
 \item{"min" }{Calculate difference from minimum of perf values - default}
 \item{"max" }{Calculate difference from maximum of perf values}
 \item{"neutral" }{Calculate difference from zero}
 }}
}
\value{
a character vector of SAS code
}
\description{
generate SAS code represenation of the Scorecard object. The SAS
code that is genereated calculates the score, adverse action code distances,
and provides a set of macro assignments for assigning adverse action codes
to particular bin levels.
}
