% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard_class.R
\name{Scorecard_get_dropped}
\alias{Scorecard_get_dropped}
\title{Return names of varibales flagged as dropped}
\value{
a character vector of dropped variable names
}
\description{
Return names of varibales flagged as dropped
}
