% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard_class.R
\name{Scorecard_get_inmodel}
\alias{Scorecard_get_inmodel}
\title{Return names of varibales flagged as inmodel}
\value{
a character vector of inmodel variable names
}
\description{
Return names of varibales flagged as inmodel
}

