% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard_class.R
\name{Scorecard_select}
\alias{Scorecard_select}
\title{Select a fitted model and load the Transforms}
\description{
Select searches the model fitting history for the requested
model. It then loads the associated transforms for the model variables into
the current Scorevard object. Additionally it loads the dropped and inmodel
vectors at the time of the fit.
}
