% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard_class.R
\name{Scorecard_sort}
\alias{Scorecard_sort}
\title{Sort variables of a scorecard}
\description{
Sort variables of a scorecard
}
\details{
variables that are in the currently selected model are sorted to the
front while dropped variables are sorted to the end. Variables within each
group are sort by descending information value.
}

