% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard_class.R
\name{Scorecard_summary}
\alias{Scorecard_summary}
\title{Summarize the currently selected model}
\arguments{
\item{keep}{whether to summarize droppped variables as well}
}
\value{
a matrix summarizing the independent variables using the performance
implementation summary function. Also displays the coefficients and model
contributions of the predictors.
}
\description{
Summarize the currently selected model
}
