% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_class.R
\name{Scorecard_undrop}
\alias{Scorecard_undrop}
\title{Flag supplied variables as undropped}
\arguments{
\item{vars}{character vector of variables to undrop}

\item{all}{logical indicating whether all variables should be undropped}
}
\description{
Flag supplied variables as undropped
}
