% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_class.R
\name{set_equal}
\alias{set_equal}
\alias{set_equal_}
\title{Set substitution of one level equal to substitution of another}
\usage{
set_equal_(tf, v1, v2)
}
\arguments{
\item{tf}{Transform object}

\item{v1}{level to override}

\item{v2}{value with which to override v1's substitution}
}
\value{
new Transform object with updated overrides
}
\description{
Set substitution of one level equal to substitution of another
}

