% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_class.R
\docType{class}
\name{Transform-class}
\alias{Transform-class}
\title{Transfrom reference class generator}
\description{
Transform object class definition. The Transform object controls
how an independent variable is binned and manages missing values, exceptions,
overrides, and weight-of-evidence substitution.
}
\section{Slots}{

\describe{
\item{\code{tf}}{the mapping of input values to bin levels}

\item{\code{subst}}{the WoE substitution for the mapped values}

\item{\code{exceptions}}{numeric vector of exception values}

\item{\code{nas}}{WoE value for missing value sof \code{x}}

\item{\code{overrides}}{named vector of bin levels to overriden substituion values}

\item{\code{repr}}{cached, tabular representation of the bin object for quick display}

\item{\code{neutralized}}{which levels are neutrazlied}
}}

