% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim}
\usage{
trim(x, contribution = 0)
}
\arguments{
\item{x}{A \code{binnr.model} or \code{segmented} object}

\item{contribution}{the contribution threshold. Variables below this
threshold will be flagged as dropped for subsequent \code{fit}
operations.}
}
\value{
an object of the same type that was provided in \code{x}. Variables
below the passed contribution threshold are now flagged as dropped.
}
\description{
Trim variables from binnr.models based on contribution
}

