% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_class.R
\name{update_transform}
\alias{update_transform}
\title{Update Transform object with new information after Bin operations}
\usage{
update_transform(tf, result)
}
\arguments{
\item{tf}{Transform object}

\item{result}{list of summarized performane values for the updated Bin object}
}
\value{
new Transform object with updated subst, nas, exceptions, and repr
}
\description{
Update Transform object with new information after Bin operations
}
