% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bin.common.R
\name{bin}
\alias{bin}
\title{Bin variables for credit modeling}
\usage{
bin(x, y, seg = NULL, name, min.iv, min.cnt, min.res, max.bin, mono,
  exceptions)
}
\arguments{
\item{x}{a \code{factor} or \code{numeric} variable. Can also be a
\code{data.frame} in which case a \code{bin.list} object is returned.}

\item{y}{a binary response variable}

\item{name}{the name used for plotting and printing. Automatically supplied
if called on a \code{data.frame}}

\item{min.iv}{minimum information value required to split a continuous
predictor}

\item{min.cnt}{minimum number of observations to split a continuous
predictor. Defaults to the square root of the number of observations.}

\item{min.res}{minimum number of responses required to split}

\item{max.bin}{maximum number of levels in resulting discretized, continuous
predictor.}

\item{mono}{monotonicity constraint for binning algorith. \code{binnr} uses
the sane assumption that the 1-labelled target is of most interest. Therefore
monotonicity of 1 implies the target rate increases with the predictor. The
opposite for -1. A value of 0 enforces no monotonicity constraint and a 2
enforces *any* monotonic constraint but doesn't assume a particular
direction. If \code{bin} is used on a \code{data.frame}, mono can be a named
vector. The reserved name ALL is used for global values.}

\item{exceptions}{a \code{list} of values to be excluded from the binning
algorithm. These levels count towards weight of evidence calculations but are
not collapsed. Exceptions values for \code{factors} are ignored. A reserved
name, ALL, can be used to supply global exception values.}
}
\value{
a \code{bin} or \code{bin.list} object
}
\description{
bin returns a \code{bin} object containing information necessary
for modeler interaction. With a \code{bin} object, a modeler can collapse,
expand, neutralize, cap, and predict variable transformations
}

