% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_class.R,
%   R/binary_performance_class.R
\docType{methods}
\name{bin_}
\alias{bin_}
\alias{bin_,Binary_Performance,Continuous-method}
\alias{bin_,Binary_Performance,Discrete-method}
\title{generic method for bin_}
\usage{
bin_(.self, b, ...)

\S4method{bin_}{Binary_Performance,Continuous}(.self, b, min.iv = 0.01,
  min.cnt = 10, min.res = 0, max.bin = 10, mono = 0,
  exceptions = numeric(0))

\S4method{bin_}{Binary_Performance,Discrete}(.self, b)
}
\arguments{
\item{.self}{the Performance object calling bin_}

\item{b}{a Bin object}

\item{min.iv}{minimum information value to split}

\item{min.cnt}{minimum number of observations per bin}

\item{min.res}{minimum number of response counts per bin}

\item{max.bin}{maximum number of resulting bins}

\item{mono}{montoncity constraint of discretization process}

\item{exceptions}{numeric vector of values to exclude from discretization#'}

\item{.self}{a Binary_Performance object}

\item{b}{a Continuous bin object}
}
\value{
modifies the Bin object in place

modifies the Bin object in place
}
\description{
generic method for bin_
}
\section{Methods (by class)}{
\itemize{
\item \code{.self = Binary_Performance,b = Continuous}: Bin a Continuous object using Binary_Performance

\item \code{.self = Binary_Performance,b = Discrete}: Bin a Discrete object using Binary_Performance
}}

