\name{binnr-package}
\alias{binnr-package}
\alias{binnr}
\docType{package}
\title{
Binnr - Scorecard development suite
}
\description{
Binnr does stuff
}
\details{
\tabular{ll}{
Package: \tab binnr\cr
Type: \tab Package\cr
Version: \tab 0.9\cr
Date: \tab 2015-09-03\cr
License: \tab Free\cr
}
Do more stuff here!
}
\author{
Eric E. Graves

Maintainer: Eric E. Graves <gravcon5@gmail.com>
}
\references{

}
\keyword{ package }
\seealso{
}
\examples{
}
