% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{inmodel}
\alias{inmodel}
\title{Visited}
\usage{
inmodel(bins)
}
\arguments{
\item{bins}{The \code{bin.object} to check for visitation}
}
\value{
logical vector the same length as bins. TRUE if the bins has been
visited with the \code{adjust} function, FALSE otherwise.
}
\description{
Function that returns T/F vector for whether a bin has been visited with
\code{adjust}
}

